Ensure Redis Custom Role cache is evicted on updating User Custom Role
