+++ src/main/java/org/chainoptim/core/user/controller/UserController.java
@@ -2,11 +2,15 @@ package org.chainoptim.core.user.controller;
 
 import org.chainoptim.core.user.dto.AssignBasicRoleDTO;
 import org.chainoptim.core.user.dto.AssignCustomRoleDTO;
-import org.chainoptim.core.user.dto.UserWithOrganizationDTO;
-import org.chainoptim.core.user.model.User;
-import org.chainoptim.core.user.service.UserService;
 import org.chainoptim.core.user.dto.UserSearchResultDTO;
+import org.chainoptim.core.user.dto.UserWithOrganizationDTO;
 import org.chainoptim.shared.search.model.PaginatedResults;
+import org.chainoptim.core.user.model.User;
+import org.chainoptim.core.user.service.CachedUserService;
+import org.chainoptim.core.user.service.UserWriteService;
+import org.chainoptim.core.user.service.UserService;
+import org.chainoptim.core.user.service.UserUpdateService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,10 +23,23 @@ import java.util.Optional;
 public class UserController {
 
     private final UserService userService;
+    private final UserUpdateService userUpdateService;
+    private final UserWriteService userWriteService;
+    private final CachedUserService cachedUserService;
+
+    private final boolean isRedisAvailable;
 
     @Autowired
-    public UserController(UserService userService) {
+    public UserController(UserService userService,
+                          UserWriteService userWriteService,
+                          UserUpdateService userUpdateService,
+                          CachedUserService cachedUserService,
+                          boolean isRedisAvailable) {
         this.userService = userService;
+        this.userWriteService = userWriteService;
+        this.userUpdateService = userUpdateService;
+        this.cachedUserService = cachedUserService;
+        this.isRedisAvailable = isRedisAvailable;
     }
 
     // Fetch
@@ -80,28 +97,38 @@ public class UserController {
     // Update
     @PutMapping("/update")
     public ResponseEntity<User> updateUser(@RequestBody User user) {
-        return ResponseEntity.ok(userService.updateUser(user));
+        return ResponseEntity.ok(userUpdateService.updateUser(user));
     }
 
     @PutMapping("/{userId}/assign-basic-role")
     public ResponseEntity<User> assignBasicRoleToUser(@PathVariable("userId") String userId, @RequestBody AssignBasicRoleDTO roleDTO) {
-        return ResponseEntity.ok(userService.assignBasicRoleToUser(userId, roleDTO.getRole()));
+        if (isRedisAvailable) {
+            System.out.println("Redis is available");
+            return ResponseEntity.ok(cachedUserService.assignBasicRoleToUser(userId, roleDTO.getRole()));
+        } else {
+            System.out.println("Redis is not available");
+            return ResponseEntity.ok(userUpdateService.assignBasicRoleToUser(userId, roleDTO.getRole()));
+        }
     }
 
     @PutMapping("/{userId}/assign-custom-role")
     public ResponseEntity<User> assignCustomRoleToUser(@PathVariable("userId") String userId, @RequestBody AssignCustomRoleDTO roleDTO) {
-        return ResponseEntity.ok(userService.assignCustomRoleToUser(userId, roleDTO.getRoleId()));
+        if (isRedisAvailable) {
+            return ResponseEntity.ok(cachedUserService.assignCustomRoleToUser(userId, roleDTO.getRoleId()));
+        } else {
+            return ResponseEntity.ok(userUpdateService.assignCustomRoleToUser(userId, roleDTO.getRoleId()));
+        }
     }
 
     @PutMapping("/{userId}/remove-from-organization/{organizationId}")
     public ResponseEntity<User> removeUserFromOrganization(@PathVariable("userId") String userId, @PathVariable("organizationId") Integer organizationId) {
-        return ResponseEntity.ok(userService.removeUserFromOrganization(userId, organizationId));
+        return ResponseEntity.ok(userUpdateService.removeUserFromOrganization(userId, organizationId));
     }
 
     // Delete
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable String id) {
-        userService.deleteUser(id);
+        userWriteService.deleteUser(id);
         return ResponseEntity.ok().build();
     }
 }
