+++ src/main/java/org/chainoptim/core/user/jwt/JwtAuthenticationFilter.java
@@ -1,6 +1,7 @@
 package org.chainoptim.core.user.jwt;
 
 import org.chainoptim.core.user.model.UserDetailsImpl;
+import org.chainoptim.core.user.service.UserDetailsServiceImpl;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -8,7 +9,6 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.chainoptim.core.user.service.UserDetailsServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -27,16 +27,16 @@ import java.util.Collections;
 @EnableWebSecurity
 public class JwtAuthenticationFilter extends GenericFilterBean {
 
-    private final JwtTokenProvider tokenProvider;
     private final UserDetailsServiceImpl userDetailsService;
+    private final JwtTokenProvider tokenProvider;
 
     @Autowired
     public JwtAuthenticationFilter(
-            JwtTokenProvider tokenProvider,
-            UserDetailsServiceImpl userDetailsService
+            UserDetailsServiceImpl userDetailsService,
+            JwtTokenProvider tokenProvider
     ) {
-        this.tokenProvider = tokenProvider;
         this.userDetailsService = userDetailsService;
+        this.tokenProvider = tokenProvider;
     }
 
     @Override
