+++ src/main/java/org/chainoptim/core/user/service/UserWriteService.java
@@ -1,0 +1,13 @@
+package org.chainoptim.core.user.service;
+
+import org.chainoptim.core.user.model.User;
+
+public interface UserWriteService {
+
+    User registerNewUser(String username, String password, String email);
+    User registerNewOrganizationUser(String username, String password, String email, Integer organizationId, User.Role role);
+
+    // Delete
+    void deleteUser(String id);
+
+}
